move logic to prepareDocs.js
